+++ src/hash/oldhat.c
@@ -192,7 +192,7 @@ oldhat_init_size(oldhat_t *self, char size)
     oldhat_store_t *store;
     uint64_t        len;
 
-    if (size > (ssize_t)(sizeof(intptr_t) * 8)) {
+    if (size > (sizeof(intptr_t) * 8)) {
         abort();
     }
