+++ src/hash/swimcap.c
@@ -106,7 +106,7 @@ swimcap_init_size(swimcap_t *self, char size)
     swimcap_store_t *store;
     uint64_t         len;
 
-    if (size > (ssize_t)(sizeof(intptr_t) * 8)) {
+    if (size > (sizeof(intptr_t) * 8)) {
         abort();
     }
